plugins/pr_config.c: Fix storing the return value of read() Use the right variable type when storing the return value of the read() function. Using an unsigned type prevents the code from catching negative return values (errors).

+++ plugins/pr_config.c
@@ -95,7 +95,7 @@ static void entry_set_hex_int(GtkWidget *entry, unsigned data)
  */
 static const char * updatePR(const char * pr_bin_path) {
 
-	uint32_t status = 0;
+	ssize_t status = 0;
 	int ret, fd;
 
 	fd = open(pr_bin_path, O_RDONLY);
